+++ src/data/index.ts
@@ -1,4 +1,7 @@
-export { default as DataClient } from "./dataClient";
+export {
+  default as DataClient,
+  type DataParams,
+} from "./dataClient";
 export { default as dataProvider } from "./dataProvider";
 
 export * from "./dataTypes";