+++ changelog.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Allow `options` props to be pass to `accessControlProvider`
 - Export `authType`
 - Properly handle `error`
+- Export `DataParams` type
 
 ## [0.0.5] - 2024-03-06